added animation next page
