+++ package.json
@@ -12,11 +12,11 @@
     "@testing-library/react": "^9.3.2",
     "@testing-library/user-event": "^7.1.2",
     "bootstrap": "^4.5.3",
+    "framer-motion": "^2.9.5",
     "get-google-fonts": "^1.2.2",
     "graphql": "^15.3.0",
     "gsap": "^3.5.1",
     "html-react-parser": "^0.14.0",
-    "prop-types": "^15.7.2",
     "react": "^16.13.1",
     "react-bootstrap": "^1.3.0",
     "react-dom": "^16.13.1",
@@ -30,7 +30,6 @@
     "react-scroll-up": "^1.3.5",
     "react-swipeable-views": "^0.13.9",
     "react-tabs": "^3.1.1",
-    "react-transition-group": "^4.4.1",
     "redux": "^4.0.5",
     "typesafe-actions": "^5.1.0"
   },
