+++ src/common/page-transition/configVarian.js
@@ -1,0 +1,10 @@
+export const pageTransition = {
+  in: {
+    opacity: 1,
+    x: 0,
+  },
+  out: {
+    opacity: 0,
+    x: "-100vw",
+  },
+};
