+++ src/components/body/home-page/Banner.js
@@ -4,9 +4,7 @@ function Banner() {
   return (
     <div id="bg-banner" className="text-center">
       <h1 className="font-weight-bold text-white">UNIVERSITY REVIEWS</h1>
-      <h3 className="text-center text-light">
-        A place to share information and feeling
-      </h3>
+      <h3 className="text-center text-light">A place to share information</h3>
     </div>
   );
 }
