+++ src/components/body/home-page/index.js
@@ -1,14 +1,16 @@
 import React, { useEffect, useState } from "react";
-import "./home-page.css";
+
 import { useQuery, useMutation } from "@apollo/client";
 import { fetchUniversityAction } from "state/ducks/common/actions/home-page";
 import { useDispatch, useSelector } from "react-redux";
 //importing local file
 import homepageQueries from "query/homepage";
 import { loginMutation } from "query/login";
+
 import CardHome from "./CardHome";
 import FilterBar from "./FilterBar";
 import Banner from "./Banner";
+import "./home-page.css";
 //importing material UI
 import Pagination from "@material-ui/lab/Pagination";
 
@@ -71,7 +73,7 @@ function Index({ match }) {
   // using callback to memoize function after each time render
 
   return (
-    <div className="body-homepage">
+    <div>
       <Banner />
       <div className="container">
         <div className="row">
