+++ src/components/body/user-profile/index.js
@@ -1,45 +1,19 @@
 import React from "react";
+import { motion } from "framer-motion";
 //importing local file
 import "./style/user.css";
 import "./style/my-post.css";
 import Header from "./Header";
 import MyPost from "./MyPost";
 import PinPost from "./PinPost";
+import TabPanel, { a11yProps } from "common/tabs/TabPanel";
+import { pageTransition } from "common/page-transition/configVarian";
 //material-ui
 import { useTheme, makeStyles } from "@material-ui/core/styles";
 import SwipeableViews from "react-swipeable-views";
 import Tabs from "@material-ui/core/Tabs";
 import Tab from "@material-ui/core/Tab";
 // import PropTypes from "prop-types";
-
-function TabPanel(props) {
-  const { children, value, index, ...other } = props;
-
-  return (
-    <div
-      role="tabpanel"
-      hidden={value !== index}
-      id={`full-width-tabpanel-${index}`}
-      aria-labelledby={`full-width-tab-${index}`}
-      {...other}
-    >
-      {value === index && <div>{children}</div>}
-    </div>
-  );
-}
-
-// TabPanel.propTypes = {
-//   children: PropTypes.node,
-//   index: PropTypes.any.isRequired,
-//   value: PropTypes.any.isRequired,
-// };
-
-function a11yProps(index) {
-  return {
-    id: `full-width-tab-${index}`,
-    "aria-controls": `full-width-tabpanel-${index}`,
-  };
-}
 
 const useStyles = makeStyles((theme) => ({
   root: {
@@ -61,13 +35,14 @@ export default function FullWidthTabs() {
   };
 
   return (
-    <div className="user">
+    <motion.div
+      initial="out"
+      animate="in"
+      exit="out"
+      variants={pageTransition}
+      className="user"
+    >
       <Header />
-      <form method="GET">
-        <label>Mien Trung</label>
-        <input type="radio" id="trung" name="zone" value="Mien Trung" />
-        <input type="submit" value="Submit" />
-      </form>
       <div className="container main-user">
         <div className={classes.root}>
           <Tabs
@@ -79,10 +54,10 @@ export default function FullWidthTabs() {
             aria-label="full width tabs example"
           >
             <Tab
-              label="Item One"
+              label="Bài viết đã đăng"
               {...a11yProps(0)} // option, it's helpful for SEO
             />
-            <Tab label="Item Two" {...a11yProps(1)} />
+            <Tab label="Bài viết yêu thích" {...a11yProps(1)} />
           </Tabs>
           <SwipeableViews
             axis={theme.direction === "rtl" ? "x-reverse" : "x"} //option, it's helpful for SEO
@@ -100,6 +75,6 @@ export default function FullWidthTabs() {
           </SwipeableViews>
         </div>
       </div>
-    </div>
+    </motion.div>
   );
 }
